Fixing package path issues and removing unwanted directories.
